+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 
+- Optional support for watchtower from [@bradrydzewski](https://github.com/bradrydzewski).
+- Optional support for drone/gc from [@bradrydzewski](https://github.com/bradrydzewski). 
 - Update the default agent image to 1.0 stable, from [@bradrydzewski](https://github.com/bradrydzewski).
 - Configure agent environment variables from [@bradrydzewski](https://github.com/bradrydzewski).
 - Configure agent host volume mounts from [@patrickjahns](https://github.com/patrickjahns).